CHORE: Auto Deploy with Travis CI (#20)
This attempts to fix https://github.com/afringer/Haku/issues/15 by
using using Travis CI to auto deploy the bot.

Currently it performs no testing as that is out of scope for this build
however that can be added in future builds. In order to auto deploy
testing returns success even if we have errors.

Configuration will need to be done after this code is merged to master
in order to configure this for production env. This include creating
a Travis-CI account and adding production env. ssh and config values.

Signed-off-by: Adam Brenner <adam@aeb.io>